requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
sentence-transformers
faiss-cpu
python-dotenv
pydantic
requests
